Added app production dependencies
- A second package.json file has been added to the app/ directory,
this file, unlike the package.json located in the root of the
project, is meant to hold the dependencies that are going to be needed
by our Electron application during runtime. While the package.json in
the root of the project is meant to hold dependencies that are necessary
only for the build process. This approach is also known as 'Two
package.json Structure'.