Implemented Redux Store modules
- Three new directories have been added to the app/ directory, each of
these directories contain modules that are necessary for the working of
our Redux store. All these modules are brought together in our
configureStore module which is responsible for assembling the Redux
Store and making sure that it can be properly integrated with our
Electron app.

- Besides the Store Actions and the Reducers, we have another element in
our Redux setup that may be perhaps a bit unconventional in a regular
Redux project, and that is the Sagas module. The Sagas are responsible
for handling not only async actions, but also any actions that result in
a side-effect. That includes mostly deleting, adding and updating
documents to our collections and keeping our Redux Store aware of these
changes in the database.