Implemented project entry files
- The main.js entry file has been implemented, this script file is used
by the Electron runtime to start the Main Process of our app, the
Main Process is, in turn, responsible for instantiating other Renderer
Process, which are just Chromium instances that have access to Node
modules.

- The Main Process is also responsible for initializing our Store, seed
the database and to configure the app settings. The Store in the Main
Process is then kept in sync with all other Stores instantiated in the
Renderer processes.

- The app.js script, along with our app.html template, is in turn going
to be used to instantiate Renderer processes.