Implemented SelectField React component
- The SelectField component has been implemented, this component
is intended to take the place of the former SelectGroup component
and be a more abstract and general component. For the time being,
it is identical to the NotebooksField sub-component, however the
NotebooksField sub-component is expect to receive modifications
that would make it more specific to the context of the NotesPage
in the future.