Implemented Panel sub-component
- The Panel sub-component of SettingsPage has been implemented.
The Panel component is fairly general container component, however
it is not expected not be used in anywhere else but the SettingsPage
for the time being. In the context of the SettingsPage, the Panel
sub-component is going to be used to group a set of related options
of the application.