Implement dynamic themes and actions on EditorPage
The EditorPage now supports dynamic themes, allowing the user to
define wich theme to use through the SettingsPage.

The Sidebar component of the EditorPage has also been rewriten to
allow the user to manipulate some aspects of the EditorPage and
the Notes collection.

Minor fixes were made to the SettingPage and the settings.json module.