Fix app's text rendering issues on windows/linux
This commit introduces changes that make use of the platform specific
css class introduced in the RootLayout to make adjustments to the text
rendering in specific platforms. This applies mainly to the TitleBars of
the various pages of the application but it can also affect other
components that have text rendering issues in specific platforms.